// src/api/models/CoddNews_Common_NewsArticle.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CoddNews_Common_NewsArticleImage } from './CoddNews_Common_NewsArticleImage';
export type CoddNews_Common_NewsArticle = {
    sourceUrl?: string | null;
    title?: string | null;
    paragraphs?: string | null;
    publishDate?: string;
    images?: Array<CoddNews_Common_NewsArticleImage> | null;
};